Add basic get view
This necessitated some refactoring around the way the templates
are constructed, there is now one top level template and the path
and parameter parts are build up before being turned into the
URL in the UserAgent methods.

The previous method of having sub-templates was suffering from
over-encoding and was always doomed to fail.